create launchableMarketsSelectors
